Updated code to allow for more minigames in the future, small games such as rock, paper, scissors, x&o's, etc Cleaned up how instructions are handled into it's own class
