+++ Guess My Number/Guess My Number/Debug/Guess My Number.Build.CppClean.log
@@ -1,15 +1,17 @@
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\vc143.pdb
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\vc143.idb
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\main.obj
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\handlefile.obj
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\highscores.obj
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\game.obj
-n:\code repositories\guessnumberrepo\guess my number\debug\guess my number.exe
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.ilk
-n:\code repositories\guessnumberrepo\guess my number\debug\guess my number.pdb
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.tlog\cl.command.1.tlog
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.tlog\cl.read.1.tlog
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.tlog\cl.write.1.tlog
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.tlog\link.command.1.tlog
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.tlog\link.read.1.tlog
-n:\code repositories\guessnumberrepo\guess my number\guess my number\debug\guess my number.tlog\link.write.1.tlog
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\vc143.pdb
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\vc143.idb
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\main.obj
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\instructions_handler.obj
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\highscores_handler.obj
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guessmynumber.obj
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\gameselection.obj
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\file_handler.obj
+n:\code repositories\guessmynumberrepo\guess my number\debug\guess my number.exe
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.ilk
+n:\code repositories\guessmynumberrepo\guess my number\debug\guess my number.pdb
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.tlog\cl.command.1.tlog
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.tlog\cl.read.1.tlog
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.tlog\cl.write.1.tlog
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.tlog\link.command.1.tlog
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.tlog\link.read.1.tlog
+n:\code repositories\guessmynumberrepo\guess my number\guess my number\debug\guess my number.tlog\link.write.1.tlog
