+++ Guess My Number/Guess My Number/Debug/Guess My Number.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>N:\Code Repositories\GuessNumberRepo\Guess My Number\Debug\Guess My Number.exe</FullPath>
+      <FullPath>N:\Code Repositories\GuessMyNumberRepo\Guess My Number\Debug\Guess My Number.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
