+++ Guess My Number/Guess My Number/Debug/Guess My Number.tlog/Guess My Number.lastbuildstate
@@ -1,2 +1,2 @@
-PlatformToolSet=v143:VCToolArchitecture=Native32Bit:VCToolsVersion=14.33.31629:TargetPlatformVersion=10.0.19041.0:
-Debug|Win32|N:\Code Repositories\GuessNumberRepo\Guess My Number\|
+PlatformToolSet=v143:VCToolArchitecture=Native32Bit:VCToolsVersion=14.34.31933:TargetPlatformVersion=10.0.22000.0:
+Debug|Win32|N:\Code Repositories\GuessMyNumberRepo\Guess My Number\|
