+++ Guess My Number/Guess My Number/Debug/HighScores.nativecodeanalysis.sarif
@@ -1,0 +1,117 @@
+{
+  "version": "2.1.0",
+  "$schema": "https://schemastore.azurewebsites.net/schemas/json/sarif-2.1.0-rtm.5.json",
+  "runs": [
+    {
+      "results": [
+        {
+          "ruleId": "C26495",
+          "message": {
+            "text": "Variable 'HighScores::HighscoreEntry::score' is uninitialized. Always initialize a member variable (type.6)."
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "artifactLocation": {
+                  "uri": "file:///N:/Code Repositories/GuessMyNumberRepo/Guess My Number/Guess My Number/HighScores.h"
+                },
+                "region": {
+                  "startLine": 21,
+                  "startColumn": 2,
+                  "endLine": 21,
+                  "endColumn": 2
+                }
+              },
+              "logicalLocations": [
+                {
+                  "fullyQualifiedName": "HighScores::HighscoreEntry::{ctor}",
+                  "decoratedName": "??0HighscoreEntry@HighScores@@QAE@XZ",
+                  "kind": "function"
+                }
+              ]
+            }
+          ],
+          "properties": {
+            "targetSymbol": "HighScores::HighscoreEntry::score"
+          }
+        },
+        {
+          "ruleId": "C26495",
+          "message": {
+            "text": "Variable 'HighScores::HighscoreEntry::position' is uninitialized. Always initialize a member variable (type.6)."
+          },
+          "locations": [
+            {
+              "physicalLocation": {
+                "artifactLocation": {
+                  "uri": "file:///N:/Code Repositories/GuessMyNumberRepo/Guess My Number/Guess My Number/HighScores.h"
+                },
+                "region": {
+                  "startLine": 21,
+                  "startColumn": 2,
+                  "endLine": 21,
+                  "endColumn": 2
+                }
+              },
+              "logicalLocations": [
+                {
+                  "fullyQualifiedName": "HighScores::HighscoreEntry::{ctor}",
+                  "decoratedName": "??0HighscoreEntry@HighScores@@QAE@XZ",
+                  "kind": "function"
+                }
+              ]
+            }
+          ],
+          "properties": {
+            "targetSymbol": "HighScores::HighscoreEntry::position"
+          }
+        }
+      ],
+      "tool": {
+        "driver": {
+          "name": "PREfast",
+          "fullName": "PREfast Code Analysis",
+          "version": "14.34.31942.0",
+          "informationUri": "https://aka.ms/cpp/ca"
+        }
+      },
+      "invocations": [
+        {
+          "executionSuccessful": true
+        }
+      ],
+      "artifacts": [
+        {
+          "location": {
+            "uri": "file:///N:/Code Repositories/GuessMyNumberRepo/Guess My Number/Guess My Number/HighScores.h"
+          },
+          "roles": [
+            "resultFile"
+          ],
+          "hashes": {
+            "sha-256": "c0e876a172a989133cd9f0834a869fbade3d558bef380224085dc2b606eb5606"
+          }
+        },
+        {
+          "location": {
+            "uri": "file:///n:/code repositories/guessmynumberrepo/guess my number/guess my number/highscores.cpp"
+          },
+          "roles": [
+            "analysisTarget"
+          ],
+          "hashes": {
+            "sha-256": "1457328a0c18e4d0eacd77ad0a6ee2b527ff53ed1ce282e1a47155c71d5bb4d8"
+          }
+        },
+        {
+          "location": {
+            "uri": "file:///n:/code repositories/guessmynumberrepo/guess my number/guess my number/handlefile.h"
+          },
+          "hashes": {
+            "sha-256": "c20aeacc6ac9343ee95041b14280f0993a0277a0235bcd3a17191e63f223a03e"
+          }
+        }
+      ]
+    }
+  ]
+}
