+++ Guess My Number/Guess My Number/GameSelection.h
@@ -1,0 +1,18 @@
+#pragma once
+#include <iostream>
+#include "GuessMyNumber.h"
+
+using namespace std;
+
+class GameSelection
+{
+private:
+	GuessMyNumber* guessNumberGame{ new GuessMyNumber() };
+
+public:
+	GameSelection() {}
+
+	void GameSelectorMenu();
+	void DisplayGamesMenu();
+};
+
