--- Guess My Number/Guess My Number/HandleFile.h
@@ -1,15 +1,0 @@
-#pragma once
-#include <string>
-#include <fstream>
-#include <iostream>
-
-using namespace std;
-
-class HandleFile
-{
-private:
-
-public:
-	int loadAll(const string& fileName, string file_lines[]);
-	int saveToFile(const string& fileName, string scores[], int size);
-};
