+++ Guess My Number/Resources/GuessMyNumber/highscores.txt
@@ -1,0 +1,10 @@
+cra1, 1
+lp2, 1
+cra444, 1
+difkjdjndndd, 1
+ffffffffffffff, 1
+ssssss, 2
+Craig1, 5
+Craig, 6
+Dave, 7
+John, 7
