--- Guess My Number/Resources/highscores.txt
@@ -1,10 +1,0 @@
-cra1, 1
-lp2, 1
-cra444, 1
-ssssss, 2
-Craig1, 5
-Craig, 6
-Dave, 7
-John, 7
-craig3, 7
-Sue, 8
